+++ src/main/java/ru/dosport/repositories/SportTypeRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import ru.dosport.entities.SportType;
 
 @Repository
-public interface SportTypeRepository extends JpaRepository<SportType,Long> {
+public interface SportTypeRepository extends JpaRepository<SportType,Short> {
 }
